requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
werkzeug
apscheduler
requests
python-dotenv
email-validator
python-docx
markdown
sendgrid